+++ test.svelte
@@ -1,18 +1,28 @@
 <script lang="ts">
+  import { onMount } from "svelte";
+
   import { Pincode, PincodeInput } from "./types";
   import { Code } from "./types/Pincode";
 
   const correctCode = "1234";
 
+  let ref: Pincode;
   let code: Code = [];
   let value = "";
 
   $: complete = value.length === correctCode.length;
   $: success = complete && value === correctCode;
   $: error = complete && !success;
+
+  onMount(() => {
+    ref.focusFirstInput();
+    ref.focusNextEmptyInput();
+    ref.focusLastInput();
+  });
 </script>
 
 <Pincode
+  bind:this="{ref}"
   bind:code
   bind:value
   on:complete="{(e) => {