+++ types/Pincode.d.ts
@@ -13,10 +13,28 @@ export interface PincodeProps extends svelte.JSX.HTMLAttributes<HTMLElementTagNa
    * @default ""
    */
   value?: string;
+
+  /**
+   * @constant
+   * @default () => { ref.querySelector("input").focus(); }
+   */
+  focusFirstInput?: () => void;
+
+  /**
+   * @constant
+   * @default () => { for (const input of ref.querySelectorAll("input")) { if (input.value.length === 0) { input.focus(); break; } } }
+   */
+  focusNextEmptyInput?: () => void;
+
+  /**
+   * @constant
+   * @default () => { ref.querySelector("input:last-of-type").focus(); }
+   */
+  focusLastInput?: () => void;
 }
 
 export default class Pincode extends SvelteComponentTyped<
   PincodeProps,
-  { complete: CustomEvent<{ code: Code }> },
+  { complete: CustomEvent<{ code: Code; value: string }> },
   { default: {} }
 > {}