+++ CHANGELOG.md
@@ -7,8 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-- `focusNextEmptyInput`
-- `focusInput`
+- Number-only variant
+
+## [0.2.0](https://github.com/metonym/svelte-pincode/releases/tag/v0.2.0) - 2021-01-01
+
+- Add `focusFirstInput`, `focusNextEmptyInput`, `focusLastInput` methods
 
 ## [0.1.0](https://github.com/metonym/svelte-pincode/releases/tag/v0.1.0) - 2021-01-01